pyproject.toml
[tool.poetry]
name = "dims"
version = "0.1.1"
description = ""
authors = ["Nina Jensen <nina.k.jensen@gmail.com>"]
license = "Unlicense"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = "^2.3.0"
pydantic = "^1.9.0"
structlog = "^21.5.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
mypy = "^0.942"
flake8 = "^4.0.1"
black = "^22.3.0"
pytest-cov = "^3.0.0"
hypothesis = "^6.43.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
follow_imports = "skip"
ignore_errors = true

[[tool.mypy.overrides]]
module = "dims.*"
disallow_untyped_defs = true
ignore_errors = false

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.1"
version_files = ["dims/__init__.py", "pyproject.toml:version"]
